Summarize the conclusion of `i64.add_wide3`
This commit is an attempt to summarize the discussion in #6 and write it
up in the overview. I've updated various sections and added a new
section about `add_wide3` as well. The tl;dr; is that the proposal is
still going to exclusively support `i64.add128` and `i64.sub128` to
defer `i64.add_wide3` to a theoretical future where `i1` is a native
WebAssembly type.

Closes #6
Closes #34